+++ backend/package.json
@@ -14,13 +14,16 @@
     "@babel/preset-env": "^7.16.11",
     "@netlify/functions": "^2.4.0",
     "@types/express": "^4.17.21",
+    "bcrypt": "^5.1.1",
     "bcrypt-nodejs": "^0.0.3",
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "express-async-handler": "^1.2.0",
     "express-list-endpoints": "^6.0.0",
     "hibp": "^14.0.0",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.0.2",
     "nodemon": "^3.0.1",
     "serverless-http": "^3.2.0"
